+++ public/js/master/categoryCodes/form.js
@@ -1,5 +1,9 @@
 
+var categoryCodeId = '';
+
 $(function () {
+    categoryCodeId = $('input[name="id"]').val();
+
     $('.btn-ok').on('click', function () {
         $(this).addClass('disabled')
             .text('processing...');
@@ -17,4 +21,33 @@ $(function () {
         });
     }
 
+    // 削除ボタン
+    $('.btn-delete').on('click', remove);
+
 });
+
+/**
+ * 削除
+ */
+function remove() {
+    if (window.confirm('元には戻せません。本当に削除しますか？')) {
+        $.ajax({
+            dataType: 'json',
+            url: '/categoryCodes/' + categoryCodeId,
+            type: 'DELETE'
+        })
+            .done(function () {
+                alert('削除しました');
+                location.href = '/categoryCodes';
+            })
+            .fail(function (jqxhr, textStatus, error) {
+                var message = '削除できませんでした';
+                if (jqxhr.responseJSON != undefined && jqxhr.responseJSON.error != undefined) {
+                    message += ': ' + jqxhr.responseJSON.error.message;
+                }
+                alert(message);
+            })
+            .always(function () {
+            });
+    }
+}