+++ CHANGELOG.md
@@ -8,7 +8,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Added
 
+- カテゴリーコード削除を追加
+
 ### Changed
+
+- update @chevre/factory
 
 ### Deprecated